requirements.txt
requests
beautifulsoup4
openai
playwright
tweepy
schedule
pyautogui
python-dotenv
pytz
